# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.User:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  service.CreateUserRequest:
    properties:
      age:
        maximum: 150
        minimum: 0
        type: integer
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  service.ListUsersResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  service.UpdateUserRequest:
    properties:
      age:
        maximum: 150
        minimum: 0
        type: integer
      email:
        type: string
      name:
        minLength: 1
        type: string
    type: object
host: localhost:8085
info:
  contact: {}
  description: CRUD API for users with dual-format Kafka event publishing (Avro +
    Protobuf).
  title: User Kafka Go API
  version: "1.0"
paths:
  /users:
    get:
      description: Get a paginated list of users
      parameters:
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ListUsersResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user and publish a CREATED event to Kafka
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a user
      tags:
      - users
  /users/{id}:
    delete:
      description: Soft-delete a user and publish a DELETED event to Kafka
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user
      tags:
      - users
    get:
      description: Get a single user by UUID
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user and publish an UPDATED event to Kafka
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a user
      tags:
      - users
schemes:
- http
swagger: "2.0"